Reorganize packages and remove inventory from barrels
