+++ src/main/kotlin/me/branchpanic/mods/stockpile/content/block/TrashCanBlock.kt
@@ -1,7 +1,7 @@
 package me.branchpanic.mods.stockpile.content.block
 
-import me.branchpanic.mods.stockpile.api.upgrade.UpgradeRegistry
 import me.branchpanic.mods.stockpile.content.blockentity.TrashCanBlockEntity
+import me.branchpanic.mods.stockpile.impl.upgrade.UpgradeRegistry
 import net.fabricmc.fabric.api.block.FabricBlockSettings
 import net.minecraft.block.*
 import net.minecraft.block.entity.BlockEntity
